Remove redundant abstract class
